requirements.txt
pandas
yfinance
google-genai
python-dotenv
requests
gunicorn
Flask
